// src/data/stack-components.d.ts
// generated by generate-stack-component-ts-types.js

export type StackComponent =
	| StackComponent_artifact_store
	| StackComponent_orchestrator
	| StackComponent_experiment_tracker
	| StackComponent_secrets_manager
	| StackComponent_container_registry
	| StackComponent_data_validator
	| StackComponent_model_deployer;

export type StackComponent_artifact_store = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "artifact_store";
	flavor: string;
	configuration: {
		authentication_secret: null;
		path: string;
		key: null;
		secret: null;
		token: null;
		client_kwargs: null;
		config_kwargs: null;
		s3_additional_kwargs: null;
	};
};

export type StackComponent_orchestrator = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "orchestrator";
	flavor: string;
	configuration: {
		synchronous: boolean;
		timeout: number;
		client_args: {};
		user_namespace: null;
		node_selectors: {};
		node_affinity: {};
		pod_settings: null;
		kubeflow_pipelines_ui_port: number;
		kubeflow_hostname: string;
		kubeflow_namespace: string;
		kubernetes_context: string;
		skip_local_validations: boolean;
		skip_cluster_provisioning: boolean;
		skip_ui_daemon_provisioning: boolean;
		labels: {};
		node_selector_constraint: null;
		service_account_path: null;
		project: string;
		location: string;
		pipeline_root: null;
		encryption_spec_key_name: null;
		workload_service_account: string;
		network: null;
		cpu_limit: null;
		memory_limit: null;
		gpu_limit: null;
	};
};

export type StackComponent_experiment_tracker = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "experiment_tracker";
	flavor: string;
	configuration: {
		experiment_name: null;
		nested: boolean;
		tags: {};
		tracking_uri: string | null;
		tracking_username: string | null;
		tracking_password: string | null;
		tracking_token: null;
		tracking_insecure_tls: boolean;
		databricks_host: null;
	};
};

export type StackComponent_secrets_manager = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "secrets_manager";
	flavor: string;
	configuration: {
		scope: string;
		namespace: null;
		project_id: string;
		region_name: string;
	};
};

export type StackComponent_container_registry = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "container_registry";
	flavor: string;
	configuration: {
		authentication_secret: null;
		uri: string;
	};
};

export type StackComponent_data_validator = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "data_validator";
	flavor: string;
	configuration: {};
};

export type StackComponent_model_deployer = {
	id: string;
	created: string;
	updated: string;
	user: string;
	project: string;
	is_shared: boolean;
	name: string;
	type: "model_deployer";
	flavor: string;
	configuration: {
		kubernetes_context: string;
		kubernetes_namespace: string;
		base_url: string;
		secret: string;
		custom_domain: null;
	};
};